requirements.txt
streamlit
langchain
google-generativeai
pymupdf
python-dotenv

